requirements.txt
scikit-image~=0.19.0
paddlepaddle-gpu~=2.2.0